requirements.txt
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
